+++ aries_cloudagent/messaging/issue_credential/v1_0/handlers/credential_stored_handler.py
@@ -34,4 +34,4 @@ class CredentialStoredHandler(BaseHandler):
 
         credential_manager = CredentialManager(context)
 
-        await credential_manager.credential_stored(context.message)
+        await credential_manager.credential_stored()