+++ demo/runners/alice.py
@@ -53,7 +53,7 @@ class AliceAgent(DemoAgent):
                 self.log("Connected")
                 self._connection_ready.set_result(True)
 
-    async def handle_aries36_v10_credentials(self, message):
+    async def handle_issue_credential(self, message):
         state = message["state"]
         credential_exchange_id = message["credential_exchange_id"]
         prev_state = self.cred_state.get(credential_exchange_id)
@@ -71,12 +71,11 @@ class AliceAgent(DemoAgent):
         if state == "offer_received":
             log_status("#15 After receiving credential offer, send credential request")
             await self.admin_POST(
-                "/aries0036/v1.0/issue_credential/"
-                f"{credential_exchange_id}/send_request"
+                "/issue-credential/records/" f"{credential_exchange_id}/send-request"
             )
 
         elif state == "stored":
-        # elif state == "credential_received": ??
+            # elif state == "credential_received": ??
             self.log("Storing credential in wallet")
             cred_id = message["credential_id"]
             log_status(f"#18.1 Stored credential {cred_id} in wallet")
@@ -90,7 +89,7 @@ class AliceAgent(DemoAgent):
             self.log("credential_definition_id", message["credential_definition_id"])
             self.log("schema_id", message["schema_id"])
 
-    async def handle_aries37_v10_presentations(self, message):
+    async def handle_present_proof(self, message):
         state = message["state"]
         presentation_exchange_id = message["presentation_exchange_id"]
         presentation_request = message["presentation_request"]
@@ -115,7 +114,7 @@ class AliceAgent(DemoAgent):
 
             # select credentials to provide for the proof
             credentials = await self.admin_GET(
-                f"/aries0037/v1.0/present_proof/{presentation_exchange_id}/credentials"
+                f"/present-proof/records/{presentation_exchange_id}/credentials"
             )
             if credentials:
                 for row in credentials:
@@ -153,8 +152,8 @@ class AliceAgent(DemoAgent):
             log_status("#26 Send the proof to X")
             await self.admin_POST(
                 (
-                    "/aries0037/v1.0/present_proof/"
-                    f"{presentation_exchange_id}/send_presentation"
+                    "/present-proof/records/"
+                    f"{presentation_exchange_id}/send-presentation"
                 ),
                 request,
             )